- accordions entfernt - fix für deprecated .live function - js gulp builder - struktur ./src angefangen - notices für leeren index in gallery angefangen
