+++ js/admin.js
@@ -1,121 +1,1 @@
-/*
- * Image Upload for Banner/Logo Link Widget
- */
-jQuery(document).ready(function($){
-    var custom_uploader;
-
-     $('body').on('click','.upload_image_button',function(e) {
-        e.preventDefault();
-        var button = $(this);
-        var id = button.attr('id').replace('_button', '');
-        var idimgid = button.attr('id').replace('url_button', 'id'); 
-        var idtitle = button.attr('id').replace('image_url_button', 'title');
-	
-	if (custom_uploader) {
-            custom_uploader.open();
-            return;
-        }
-        //Extend the wp.media object
-        custom_uploader = wp.media.frames.file_frame = wp.media({
-            title: 'Choose Image',
-            button: { text:'Choose Image' },
-            library: { type: 'image' }, 
-            multiple: false
-        });
- 
-        //When a file is selected, grab the URL and set it as the text field's value
-        custom_uploader.on('select', function() {
-            var attachment = custom_uploader.state().get('selection').first().toJSON();            
-            $('#'+id).val(attachment.url);   
-            $('#'+idimgid).val(attachment.id);   
-            var pretitle = $('#'+idtitle).val();
-            if (!pretitle)
-                $('#'+idtitle).val(attachment.title);
-           
-        });
- 
-        //Open the uploader dialog
-        custom_uploader.open(); 
-    });
-
-   
-    if (($('#page_template').val() == 'page-templates/page-start.php')
-     || ($('#page_template').val() == 'page-templates/page-start-sub.php')
-     || ($('#page_template').val() == 'page-templates/page-portalindex.php')
-     || ($('#page_template').val() == 'page-templates/page-portal.php')) {
-	// show the meta box
-	$('#fau_metabox_page_imagelinks').show();
-	$('#fau_metabox_page_portalmenu').show();
-	$('#fau_metabox_page_subnavmenu').hide();
-	
-
-	if (($('#page_template').val() == 'page-templates/page-portal.php')
-  	 || ($('#page_template').val() == 'page-templates/page-portalindex.php')) {
-	    $('#portalseitenquote').show();
-	    $('#fau_metabox_page_untertitel').show();
-	} else {
-	    $('#portalseitenquote').hide();
-	    $('#fau_metabox_page_untertitel').hide();
-	    
-	}
-	
-    } else {
-	// hide your meta box
-	$('#fau_metabox_page_imagelinks').hide();
-	$('#fau_metabox_page_portalmenu').hide();
-	$('#fau_metabox_page_untertitel').show();
-	 
-	if ($('#page_template').val() == 'page-templates/page-subnav.php') {
-	    // show the meta box
-	    $('#fau_metabox_page_subnavmenu').show();
-
-	} else {
-	    // hide your meta box
-	    $('#fau_metabox_page_subnavmenu').hide();
-        }
-
-    }
-
-
-    $('#page_template').live('change', function(){
-	if (($(this).val() == 'page-templates/page-start.php') 
-   	        || ($(this).val() == 'page-templates/page-start-sub.php')
-		|| ($(this).val() == 'page-templates/page-portal.php')
-		|| ($(this).val() == 'page-templates/page-portalindex.php')) {
-	    // show the meta box
-	    $('#fau_metabox_page_imagelinks').show();
-	    $('#fau_metabox_page_subnavmenu').hide();   
-	    $('#fau_metabox_page_portalmenu').show();
-
-	    if (($(this).val() == 'page-templates/page-portal.php')
-	      || ($(this).val() == 'page-templates/page-portalindex.php')) {
-		$('#portalseitenquote').show();
-		$('#fau_metabox_page_untertitel').show();
-	    } else {
-		$('#portalseitenquote').hide();
-		$('#fau_metabox_page_untertitel').hide();
-	    }
-
-	   
-	   
-	} else {
-	    // hide your meta box
-	    $('#fau_metabox_page_imagelinks').hide();
-	    $('#fau_metabox_page_portalmenu').hide();
-	    $('#fau_metabox_page_untertitel').show();
-	    if ($(this).val() == 'page-templates/page-subnav.php') {
-		// show the meta box
-		$('#fau_metabox_page_subnavmenu').show();
-
-	    } else {
-		// hide your meta box
-		$('#fau_metabox_page_subnavmenu').hide();
-
-	    }
-	}
-    });
-    
-
-   
-
-});
+"use strict";jQuery(document).ready(function(o){var _;o("body").on("click",".upload_image_button",function(e){e.preventDefault();var a=o(this),t=a.attr("id").replace("_button",""),p=a.attr("id").replace("url_button","id"),l=a.attr("id").replace("image_url_button","title");_||(_=wp.media.frames.file_frame=wp.media({title:"Choose Image",button:{text:"Choose Image"},library:{type:"image"},multiple:!1})).on("select",function(){var e=_.state().get("selection").first().toJSON();o("#"+t).val(e.url),o("#"+p).val(e.id),o("#"+l).val()||o("#"+l).val(e.title)}),_.open()}),"page-templates/page-start.php"==o("#page_template").val()||"page-templates/page-start-sub.php"==o("#page_template").val()||"page-templates/page-portalindex.php"==o("#page_template").val()||"page-templates/page-portal.php"==o("#page_template").val()?(o("#fau_metabox_page_imagelinks").show(),o("#fau_metabox_page_portalmenu").show(),o("#fau_metabox_page_subnavmenu").hide(),"page-templates/page-portal.php"==o("#page_template").val()||"page-templates/page-portalindex.php"==o("#page_template").val()?(o("#portalseitenquote").show(),o("#fau_metabox_page_untertitel").show()):(o("#portalseitenquote").hide(),o("#fau_metabox_page_untertitel").hide())):(o("#fau_metabox_page_imagelinks").hide(),o("#fau_metabox_page_portalmenu").hide(),o("#fau_metabox_page_untertitel").show(),"page-templates/page-subnav.php"==o("#page_template").val()?o("#fau_metabox_page_subnavmenu").show():o("#fau_metabox_page_subnavmenu").hide()),o(document).on("change","#page_template",function(){"page-templates/page-start.php"==o(this).val()||"page-templates/page-start-sub.php"==o(this).val()||"page-templates/page-portal.php"==o(this).val()||"page-templates/page-portalindex.php"==o(this).val()?(o("#fau_metabox_page_imagelinks").show(),o("#fau_metabox_page_subnavmenu").hide(),o("#fau_metabox_page_portalmenu").show(),"page-templates/page-portal.php"==o(this).val()||"page-templates/page-portalindex.php"==o(this).val()?(o("#portalseitenquote").show(),o("#fau_metabox_page_untertitel").show()):(o("#portalseitenquote").hide(),o("#fau_metabox_page_untertitel").hide())):(o("#fau_metabox_page_imagelinks").hide(),o("#fau_metabox_page_portalmenu").hide(),o("#fau_metabox_page_untertitel").show(),"page-templates/page-subnav.php"==o(this).val()?o("#fau_metabox_page_subnavmenu").show():o("#fau_metabox_page_subnavmenu").hide())})});
