+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "FAU-Einrichtungen",
   "textdomain": "fau",
-  "version": "1.12.14",
+  "version": "1.12.15",
   "keywords": [
     "WordPress",
     "Theme",
@@ -68,22 +68,24 @@
     "pot": "gulp pot",
     "clone": "gulp clone"
   },
+  "jsdir": "./js/",
+  "cssdir": "./css/",
   "devDependencies": {
-    "@babel/core": "^7.11.0",
+    "@babel/core": "^7.11.1",
     "@babel/preset-env": "^7.11.0",
     "@babel/register": "^7.10.5",
     "autoprefixer": "^9.8.6",
     "cssnano": "^4.1.10",
     "gulp": "^4.0.2",
+    "gulp-babel": "^8.0.0",
     "gulp-bump": "^3.2.0",
-    "gulp-concat": "^2.6.1",
     "gulp-dart-sass": "^0.9.1",
     "gulp-header": "^2.0.9",
     "gulp-postcss": "^8.0.0",
     "gulp-rename": "^1.4.0",
     "gulp-replace": "^1.0.0",
-    "gulp-sourcemaps": "^2.6.5",
     "gulp-touch-cmd": "0.0.1",
+    "gulp-uglify": "^3.0.2",
     "gulp-w3c-css": "^2.0.0",
     "gulp-wp-pot": "^2.4.3",
     "map-stream": "0.0.7",
