+++ style.css
@@ -1,7 +1,7 @@
 @charset "UTF-8";
 /*!
 Theme Name: FAU-Einrichtungen
-Version: 1.12.14
+Version: 1.12.15
 Description: WordPress-Theme der Friedrich-Alexander-Universität Erlangen-Nürnberg (FAU)
 Theme URI: https://github.com/RRZE-Webteam/FAU-Einrichtungen
 GitHub Theme URI: https://github.com/RRZE-Webteam/FAU-Einrichtungen
